+++ test/simplification/simplify.spec.ts
@@ -17,34 +17,4 @@ describe('Simplification', function() {
     expect(actualModels).toEqual(expectedModels);
     expect(schema.$id).toBeUndefined();
   });
-  test('should support simple extend', function() {
-    const inputSchemaString = fs.readFileSync(path.resolve(__dirname, './simplify/extend.json'), 'utf8');
-    const expectedSchemaString = fs.readFileSync(path.resolve(__dirname, './simplify/expected/extend.json'), 'utf8');
-    const schema = JSON.parse(inputSchemaString);
-    const expectedModels = JSON.parse(expectedSchemaString);
-    const simplifier = new Simplifier();
-    const actualModels = simplifier.simplify(schema);
-    expect(actualModels).toEqual(expectedModels);
-    expect(schema.$id).toBeUndefined();
-  });
-  test('should support advanced extend with nested objects', function() {
-    const inputSchemaString = fs.readFileSync(path.resolve(__dirname, './simplify/extendMultipleObjects.json'), 'utf8');
-    const expectedSchemaString = fs.readFileSync(path.resolve(__dirname, './simplify/expected/extendMultipleObjects.json'), 'utf8');
-    const schema = JSON.parse(inputSchemaString);
-    const expectedModels = JSON.parse(expectedSchemaString);
-    const simplifier = new Simplifier();
-    const actualModels = simplifier.simplify(schema);
-    expect(actualModels).toEqual(expectedModels);
-    expect(schema.$id).toBeUndefined();
-  });
-  test('should support advanced extend with nested objects', function() {
-    const inputSchemaString = fs.readFileSync(path.resolve(__dirname, './simplify/extendWithProperties.json'), 'utf8');
-    const expectedSchemaString = fs.readFileSync(path.resolve(__dirname, './simplify/expected/extendWithProperties.json'), 'utf8');
-    const schema = JSON.parse(inputSchemaString);
-    const expectedModels = JSON.parse(expectedSchemaString);
-    const simplifier = new Simplifier();
-    const actualModels = simplifier.simplify(schema);
-    expect(actualModels).toEqual(expectedModels);
-    expect(schema.$id).toBeUndefined();
-  });
 });
